backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
Pillow>=10.0.0
cryptography>=41.0.7
PyQt6>=6.6.0
pyinstaller>=6.0.0
